+++ src/main/java/roomescape/service/dto/request/PaymentRequest.java
@@ -1,4 +1,6 @@
 package roomescape.service.dto.request;
 
-public record PaymentRequest(String paymentKey, String orderId, int amount) {
+import java.math.BigDecimal;
+
+public record PaymentRequest(String paymentKey, String orderId, BigDecimal amount) {
 }
